returns the number of bits you would need to flip

+++ bit_manipulation/5-flip_bits.c
@@ -7,9 +7,9 @@
  * Description: Function that count set bits.
  * Return: count.
  */
-int countSetBits(int n)
+int countSetBits(unsigned long int n)
 {
-	int count = 0;
+	unsigned long int count = 0;
 
 	while (n > 0)
 	{
